feat(application status): add status expandable functionality (#811)
https://github.com/eclipse-tractusx/portal-frontend/issues/803
